Create HappyNumber.java
This solution works on the belief that if a number is not happy number, it will keep getting reduced anyway. That is, it will not be a large number.
Let us say, the number entered is a large number -> say of 9 digits -> even then -> the square of each digit will be a 2 digit number -> The sum will be a sum of 9 two-digit numbers -> which is still a small number. 

By hit and trial you will see that the numbers keep repeating every now and then. So it is rather easy to break out of the loop.